Add files via upload
there is the manual bridge code written in embedded C language , which will connect 4 Stepper Motors , and the printed parts of the bridge, will run , detecting some entities (ex. - ship, boat) by Ultrasonic Sensor, and will open the Bridge, And whenever the Object Crosses then the Bridge will close. and there is the web representation of our working process, whenever we want to control it Manually , we will use the web to control it, which is having been designed with some buttons by Css , through which we can control the Mechanism.